Tuning parameter selection information criteria

Provide information criteria for selecting tuning parameters.

Usage

criterion(hatOmega, S, n, crit, ebic.tuning = 0.5)

Arguments

hatOmega

The estimated precision matrix.

S

The sample covariance matrix.

n

An integer specifying the sample size.

crit

A string specifying the tuning parameter selection criterion to use. Available options include:

"AIC": Akaike information criterion akaike1973informationfcstat.

"BIC": Bayesian information criterion schwarz1978estimatingfcstat.

"EBIC": extended Bayesian information criterion foygel2010extendedfcstat.

"HBIC": high dimensional Bayesian information criterion wang2013calibrating,fan2017highfcstat.

ebic.tuning

A scalar (default = 0.5) specifying the tuning parameter to calculate for crit = "EBIC".

Value

A scalar.

References