Statistical methods for estimating precision matrix

Provide a collection of statistical methods to estimate a precision matrix.

Usage

fcstat(
  X,
  method,
  base = "cov",
  approach = "smp",
  lambda = NULL,
  nlambda = 50,
  lambda.min.ratio = NULL,
  lambda.min = NULL,
  lambda.max = NULL,
  gamma = NA,
  target = 0,
  initial = "glasso",
  crit = "BIC",
  ebic.tuning = 0.5,
  scaling.ratio = 0.5
)

Arguments

X

An n-by-p data matrix with sample size n and dimension p.

method

A character string specifying the statistical methods for estimating precision matrix. Available options include:

"glasso": graphical lasso

"ridge": graphical ridge

"elnet": graphical elastic net

"clime": constrained L1-minimization for inverse (covariance) matrix estimation

"tiger": tuning-insensitive graph estimation and regression

"adapt": adaptive lasso

"atan": arctangent type penalty

"exp": exponential type penalty

"mcp": minimax concave penalty

"scad": smoothly clipped absolute deviation

base

A character string (default = "cov") specifying the calculation base, either the covariance matrix ("cov") or the correlation matrix ("cor").

approach

A character string (default = "smp") specifying the approach used to compute the calculation base. Available options include:

"smp": sample (conventional approach).

"lin": linear shrinkage.

"nlminb": non-linear shrinkage using the optimization routine "nlminb".

"nloptr": non-linear shrinkage using the optimization routine "nloptr".

lambda

Grid of non-negative scalars for the regularization parameter. The default is NULL, which generates its own lambda sequence based on nlambda and lambda.min.ratio. For method = "clime", the lambda sequence is based on nlambda, lambda.min and lambda.max.

nlambda

An integer (default = 50) specifying the number of lambda values to be generated when lambda = NULL.

lambda.min.ratio

A scalar specifying the fraction of the maximum lambda value \(\lambda_{max}\) to generate the minimum lambda \(\lambda_{min}\). If lambda = NULL, the program automatically generates a lambda grid as a sequence of length nlambda in log scale, starting from \(\lambda_{min}\) to \(\lambda_{max}\), expect for method = "clime". The default value is 0.4 for method = "tiger" and 0.01 for others.

lambda.min

A scalar specifying the minimum value of program generated lambda grid for method = "clime". Default is 1e-4 (\(n>p\)) or 1e-2 (\(n<p\)).

lambda.max

A scalar (default = 0.8) specifying the maximum value of program generated lambda grid for method = "clime".

gamma

A scalar specifying the hyperparameter for the chosen method. Default values:

"elnet": a sequence from 0.1 to 0.9 with increments of 0.1

"adapt": 0.5

"atan": 0.005

"exp": 0.01

"scad": 3.7

"mcp": 3

target

A p-by-p symmetric matrix or a scalar (default = 0) serves as the value for all elements, specifying the target matrix for method = "ridge" or method = "elnet".

initial

A character string (default = "glasso") specifying

The initial estimate for method set to "atan", "exp", "scad", or "mcp".

"glasso": use the precision matrix estimate derived from the graphical lasso.

"Fan": use the inverse calculation base matrix for \(p < n\); use the precision matrix estimate derived from the graphical lasso for \(p \geq n\).

The adaptive weight for method = "adapt", calculated as \(w_{ij} = |\tilde{w}_{ij}|^{-\gamma}\), where \(\tilde{\Omega} := (\tilde{w}_{ij})\).

"glasso": use the precision matrix estimate derived from the graphical lasso as \(\tilde{\Omega}\).

"Fan": use the inverse calculation base matrix as \(\tilde{\Omega}\) for \(p < n\); use the precision matrix estimate derived from the graphical lasso as \(\tilde{\Omega}\) for \(p \geq n\).

crit

A string specifying the tuning parameter selection criterion to use. Available options include:

"AIC": Akaike information criterion (Akaike 1973) .

"BIC": Bayesian information criterion (Schwarz 1978) .

"EBIC": extended Bayesian information criterion (Foygel and Drton 2010) .

"HBIC": high dimensional Bayesian information criterion (Wang et al. 2013; Fan et al. 2017) .

ebic.tuning

A scalar (default = 0.5) specifying the tuning parameter to calculate when crit = "EBIC".

scaling.ratio

A scalar (default = 0.5) specifying the scaling ratio to adjust the lambda grid by multiplying the minimum and maximum with the scaling ratio for the case where the optimal regularization parameter lambda is found at the first index of lambda grid, make the process run iteratively with the expanded grid until the optimal parameter is no longer at the first index. process with the adjusted expanded lambda grid until the optimal regularization parameter lambda is not the first index.

Value

A list containing the following components:

hatOmega_opt

Estimated precision matrix.

lambda_opt

Optimal regularization parameter.

gamma_opt

Optimal hyperparameter.

score_opt

Information criterion score.

lambda

Actual lambda grid used in the program.

gamma

Actual gamma grid used in the program.

References

Akaike H (1973). “Information Theory and an Extension of the Maximum Likelihood Principle.” In Petrov BN, Csáki F (eds.), Second International Symposium on Information Theory, 267--281. Akad\'emiai Kiad\'o, Budapest, Hungary.

Fan J, Liu H, Ning Y, Zou H (2017). “High Dimensional Semiparametric Latent Graphical Model for Mixed Data.” Journal of the Royal Statistical Society Series B: Statistical Methodology, 79(2), 405--421. https://doi.org/10.1111/rssb.12168.

Foygel R, Drton M (2010). “Extended Bayesian Information Criteria for Gaussian Graphical Models.” In Lafferty J, Williams C, Shawe-Taylor J, Zemel R, Culotta A (eds.), Advances in Neural Information Processing Systems 23 (NIPS 2010), 604--612.

Schwarz G (1978). “Estimating the Dimension of a Model.” The Annals of Statistics, 6(2), 461--464. https://doi.org/10.1214/aos/1176344136.

Wang L, Kim Y, Li R (2013). “Calibrating Nonconvex Penalized Regression in Ultra-High Dimension.” The Annals of Statistics, 41(5), 2505--2536. https://doi.org/10.1214/13-AOS1159.