Estimation for precision matrix

Provide a collection of statistical methods to obtain a series of precision matrix estimates across the grid of tuning parameters.

Usage

fcstat.est(
  X,
  method,
  base = "cov",
  approach = "smp",
  lambda = NULL,
  nlambda = 50,
  lambda.min.ratio = NULL,
  lambda.min = NULL,
  lambda.max = NULL,
  gamma = NA,
  target = 0,
  initial = "glasso"
)

Arguments

X

An n-by-p data matrix with sample size n and dimension p.

method

A character string specifying the statistical method for estimating precision matrix. Available options include:

"glasso": graphical lasso (Friedman et al. 2008) .

"ridge": graphical ridge (van Wieringen and Peeters 2016) .

"elnet": graphical elastic net (Zou and Hastie 2005) .

"clime": constrained L1-minimization for inverse (covariance) matrix estimation (Cai et al. 2011) .

"tiger": tuning-insensitive graph estimation and regression (Liu and Wang 2017) .

"adapt": adaptive lasso (Zou 2006; Fan et al. 2009) .

"atan": arctangent type penalty (Wang and Zhu 2016) .

"exp": exponential type penalty (Wang et al. 2018) .

"mcp": minimax concave penalty (Zou 2006) .

"scad": smoothly clipped absolute deviation (Fan and Li 2001; Fan et al. 2009) .

base

A character string (default = "cov") specifying the calculation base, either the covariance matrix ("cov") or the correlation matrix ("cor").

approach

A character string (default = "smp") specifying the approach used to compute the calculation base. Available options include:

"smp": sample (conventional approach).

"lin": linear shrinkage (Ledoit and Wolf 2004) .

"nlminb": non-linear shrinkage using the optimization routine "nlminb" (Ledoit and Wolf 2015; Ledoit and Wolf 2017) .

"nloptr": non-linear shrinkage using the optimization routine "nloptr" (Ledoit and Wolf 2015; Ledoit and Wolf 2017) .

lambda

Grid of non-negative scalars for the regularization parameter. The default is NULL, which generates its own lambda sequence based on nlambda and lambda.min.ratio. For method = "clime", the lambda sequence is based on nlambda, lambda.min and lambda.max.

nlambda

An integer (default = 50) specifying the number of lambda values to be generated when lambda = NULL.

lambda.min.ratio

A scalar specifying the fraction of the maximum lambda value \(\lambda_{max}\) to generate the minimum lambda \(\lambda_{min}\). If lambda = NULL, the program automatically generates a lambda grid as a sequence of length nlambda in log scale, starting from \(\lambda_{min}\) to \(\lambda_{max}\), expect for method = "clime". The default value is 0.4 for method = "tiger" and 0.01 for others.

lambda.min

A scalar specifying the minimum value of program generated lambda grid for method = "clime". Default is 1e-4 (\(n>p\)) or 1e-2 (\(n<p\)).

lambda.max

A scalar (default = 0.8) specifying the maximum value of program generated lambda grid for method = "clime".

gamma

A scalar specifying the hyperparameter for the chosen method. Default values:

"elnet": a sequence from 0.1 to 0.9 with increments of 0.1

"adapt": 0.5

"atan": 0.005

"exp": 0.01

"scad": 3.7

"mcp": 3

target

A p-by-p symmetric matrix or a scalar (default = 0) serves as the value for all elements, specifying the target matrix for method = "ridge" or method = "elnet".

initial

A character string (default = "glasso") specifying

The initial estimate for method set to "atan", "exp", "scad", or "mcp".

"glasso": use the precision matrix estimate derived from the graphical lasso.

"Fan": use the inverse calculation base matrix for \(p < n\); use the precision matrix estimate derived from the graphical lasso for \(p \geq n\).

The adaptive weight for method = "adapt", calculated as \(|\tilde{\omega}_{ij}|^{-\gamma}\), where \(\tilde{\Omega} := (\tilde{\omega}_{ij})\).

"glasso": use the precision matrix estimate derived from the graphical lasso as \(\tilde{\Omega}\).

"Fan": use the inverse calculation base matrix as \(\tilde{\Omega}\) for \(p < n\); use the precision matrix estimate derived from the graphical lasso as \(\tilde{\Omega}\) for \(p \geq n\).

Value

An object with S3 class "fcstat.est" containing the following components:

hatOmega

A list of estimated precision matrices for lambda grid and gamma grid.

X

The raw n-by-p data matrix.

method

The statistical method used for estimating precision matrix.

base

The calculation base.

approach

The approach used to compute the calculation base.

lambda

The actual lambda grid used in the program, corresponding to hatOmega.

gamma

The actual gamma grid used in the program, corresponding to hatOmega.

target

The target matrix.

intial

The initial estimate or the adaptie weight.

Note

For the method tiger, the estimation process solely relies on the raw n-by-p data X and does not utilize the arguments base and approach. These arguments are not applicable for tiger and will have no effect if provided.

References

Cai TT, Liu W, Luo X (2011). “A Constrained \(\ell\)1 Minimization Approach to Sparse Precision Matrix Estimation.” Journal of the American Statistical Association, 106(494), 594--607. doi:10.1198/jasa.2011.tm10155 .

Fan J, Feng Y, Wu Y (2009). “Network Exploration via the Adaptive LASSO and SCAD Penalties.” The Annals of Applied Statistics, 3(2), 521--541. doi:10.1214/08-aoas215 .

Fan J, Li R (2001). “Variable Selection via Nonconcave Penalized Likelihood and its Oracle Properties.” Journal of the American Statistical Association, 96(456), 1348--1360. doi:10.1198/016214501753382273 .

Friedman J, Hastie T, Tibshirani R (2008). “Sparse Inverse Covariance Estimation with the Graphical Lasso.” Biostatistics, 9(3), 432--441. doi:10.1093/biostatistics/kxm045 .

Ledoit O, Wolf M (2004). “A Well-Conditioned Estimator for Large-Dimensional Covariance Matrices.” Journal of Multivariate Analysis, 88(2), 365--411. doi:10.1016/S0047-259X(03)00096-4 .

Ledoit O, Wolf M (2015). “Spectrum Estimation: A Unified Framework for Covariance Matrix Estimation and PCA in Large Dimensions.” Journal of Multivariate Analysis, 139, 360--384. doi:10.1016/j.jmva.2015.04.006 .

Ledoit O, Wolf M (2017). “Numerical Implementation of the QuEST Function.” Computational Statistics & Data Analysis, 115, 199--223. doi:10.1016/j.csda.2017.06.004 .

Liu H, Wang L (2017). “TIGER: A Tuning-Insensitive Approach for Optimally Estimating Gaussian Graphical Models.” Electronic Journal of Statistics, 11(1), 241--294. doi:10.1214/16-EJS1195 .

van Wieringen WN, Peeters CFW (2016). “Ridge Estimation of Inverse Covariance Matrices from High-Dimensional Data.” Computational Statistics & Data Analysis, 103, 284--303. doi:10.1016/j.csda.2016.05.012 .

Wang Y, Fan Q, Zhu L (2018). “Variable Selection and Estimation using a Continuous Approximation to the \(L_0\) Penalty.” Annals of the Institute of Statistical Mathematics, 70(1), 191--214. doi:10.1007/s10463-016-0588-3 .

Wang Y, Zhu L (2016). “Variable Selection and Parameter Estimation with the Atan Regularization Method.” Journal of Probability and Statistics, 2016, 6495417. doi:10.1155/2016/6495417 .

Zou H (2006). “The Adaptive Lasso and Its Oracle Properties.” Journal of the American Statistical Association, 101(476), 1418--1429. doi:10.1198/016214506000000735 .

Zou H, Hastie T (2005). “Regularization and Variable Selection via the Elastic Net.” Journal of the Royal Statistical Society Series B: Statistical Methodology, 67(2), 301--320. doi:10.1111/j.1467-9868.2005.00527.x .