Selection for precision matrix estimates

Provide a collection of methods to perform parameter selection for fcstat.est object.

Usage

fcstat.sel(est.obj, crit = "CV", fold = 5, ebic.tuning = 0.5)

Arguments

est.obj

The fcstat.est object output from fcstat.est().

crit

A string (default = "CV") specifying the parameter selection method to use. Available options include:

"AIC": Akaike information criterion (Akaike 1973) .

"BIC": Bayesian information criterion (Schwarz 1978) .

"EBIC": extended Bayesian information criterion (Foygel and Drton 2010) .

"HBIC": high dimensional Bayesian information criterion (Wang et al. 2013; Fan et al. 2017) .

"CV": k-fold cross validation with negative log-likelihood loss.

fold

An integer (default = 5) specifying the number of folds used for crit = "CV".

ebic.tuning

A scalar (default = 0.5) specifying the tuning parameter to calculate for crit = "EBIC".

Value

For crit = "CV", an object with S3 class "fcstat.sel" containing the following components:

hatOmega_opt

The estimated precision matrix.

lambda_opt

The optimal regularization parameter.

gamma_opt

The optimal hyperparameter.

loss_opt

The optimal k-fold loss.

lambda

The actual lambda grid used in the program.

gamma

The actual gamma grid used in the program.

loss.mean

The mean of k-fold loss for each parameter grid value.

loss.sd

The standard deviation of k-fold loss for each parameter grid value.

For other criteria, an object with S3 class "fcstat.sel" containing the following components:

hatOmega_opt

The estimated precision matrix.

lambda_opt

The optimal regularization parameter.

gamma_opt

The optimal hyperparameter.

score_opt

The optimal information criterion score.

lambda

The actual lambda grid used in the program.

gamma

The actual gamma grid used in the program.

score

The information criterion score for each parameter grid value.

References

Akaike H (1973). “Information Theory and an Extension of the Maximum Likelihood Principle.” In Petrov BN, Csáki F (eds.), Second International Symposium on Information Theory, 267--281. Akad\'emiai Kiad\'o, Budapest, Hungary.

Fan J, Liu H, Ning Y, Zou H (2017). “High Dimensional Semiparametric Latent Graphical Model for Mixed Data.” Journal of the Royal Statistical Society Series B: Statistical Methodology, 79(2), 405--421. doi:10.1111/rssb.12168 .

Foygel R, Drton M (2010). “Extended Bayesian Information Criteria for Gaussian Graphical Models.” In Lafferty J, Williams C, Shawe-Taylor J, Zemel R, Culotta A (eds.), Advances in Neural Information Processing Systems 23 (NIPS 2010), 604--612.

Schwarz G (1978). “Estimating the Dimension of a Model.” The Annals of Statistics, 6(2), 461--464. doi:10.1214/aos/1176344136 .

Wang L, Kim Y, Li R (2013). “Calibrating Nonconvex Penalized Regression in Ultra-High Dimension.” The Annals of Statistics, 41(5), 2505--2536. doi:10.1214/13-AOS1159 .