Ledoit-Wolf shrinkage estimator

Compute the Ledoit-Wolf shrinkage estimator for the covariance or correlation matrix.

Usage

ledoit_wolf_est(X, method = "lin", res = "cov")

Arguments

X

A data matrix.

method

A character string (default = "lin") specifying the method used in shrinkage, includes:

Linear shrinkage: "lin" (Ledoit and Wolf 2004) .

Non-linear shrinkage (Ledoit and Wolf 2015; Ledoit and Wolf 2017) :

"nlminb": non-linear shrinkage using the optimization routine "nlminb".

"nloptr": non-linear shrinkage using the optimization routine "nloptr".

See nlminb and nloptr for details.

res

A character string (default = "cov") specifying the result matrix to be obtained, either the covariance matrix ("cov") or the correlation matrix ("cor").

Value

A numeric matrix.

References

Ledoit O, Wolf M (2004). “A Well-Conditioned Estimator for Large-Dimensional Covariance Matrices.” Journal of Multivariate Analysis, 88(2), 365--411. doi:10.1016/S0047-259X(03)00096-4 .

Ledoit O, Wolf M (2015). “Spectrum Estimation: A Unified Framework for Covariance Matrix Estimation and PCA in Large Dimensions.” Journal of Multivariate Analysis, 139, 360--384. doi:10.1016/j.jmva.2015.04.006 .

Ledoit O, Wolf M (2017). “Numerical Implementation of the QuEST Function.” Computational Statistics & Data Analysis, 115, 199--223. doi:10.1016/j.csda.2017.06.004 .